+++ lib/danger/commands/local.rb
@@ -48,7 +48,8 @@ module Danger
       dm.env.scm = GitRepo.new
 
       dm.env.scm.diff_for_folder(".", from: dm.env.ci_source.base_commit, to: dm.env.ci_source.head_commit)
-      dm.parse Pathname.new(@dangerfile_path)
+      dm.parse(Pathname.new(@dangerfile_path))
+      dm.print_results
     end
   end
 end