+++ lib/danger/dangerfile.rb
@@ -90,5 +90,34 @@ module Danger
         # rubocop:enable Lint/RescueException
       end
     end
+
+    def print_results
+      return if (self.errors + self.warnings + self.messages + self.markdowns).count == 0
+
+      puts ""
+      puts "danger results:"
+      [:errors, :warnings, :messages].each do |current|
+        params = {}
+        params[:rows] = self.send(current).collect { |a| [a.message] }
+        next unless params[:rows].count > 0
+        params[:title] = case current
+                         when :errors
+                           current.to_s.capitalize.red
+                         when :warnings
+                           current.to_s.capitalize.yellow
+                         else
+                           current.to_s.capitalize
+                         end
+
+        puts ""
+        puts Terminal::Table.new(params)
+        puts ""
+      end
+
+      puts "Markdown: ".green if self.markdowns.count > 0
+      self.markdowns.each do |current_markdown|
+        puts current_markdown
+      end
+    end
   end
 end