+++ lib/danger/dangerfile_dsl.rb
@@ -16,12 +16,46 @@ module Danger
         self.errors = []
         self.messages = []
         self.markdowns = []
-        load_plugins
+
+        load_default_plugins
       end
 
-      def load_plugins
-        Dir['./lib/danger/plugins/*.rb'].each do |file|
-          require file
+      def import(path)
+        raise "`import` requires a string" unless path.kind_of?(String)
+        path += ".rb" unless path.end_with?(".rb")
+
+        if path.start_with?("http")
+          import_url(path)
+        else
+          import_local(path)
+        end
+      end
+
+      # Download a remote plugin and use it locally
+      #
+      # @param    [String] url
+      #           https URL to the Ruby file to use
+      def import_url(url)
+        raise "URL is not https, for security reasons `danger` only supports encrypted requests" unless url.start_with?("https://")
+
+        require 'tmpdir'
+        require 'faraday'
+        content = Faraday.get(url)
+        Dir.mktmpdir do |dir|
+          path = File.join(dir, "temporary_remote_action.rb")
+          File.write(path, content.body)
+          import_local(path)
+        end
+      end
+
+      # Import one or more local plugins
+      #
+      # @param    [String] path
+      #           The path to the file to import
+      #           Can also be a pattern (./**/*plugin.rb)
+      def import_local(path)
+        Dir[path].each do |file|
+          require File.expand_path(file) # without the expand_path it would fail if the path doesn't start with ./
         end
       end
 
@@ -76,18 +110,42 @@ module Danger
       # When an undefined method is called, we check to see if it's something
       # that either the `scm` or the `request_source` can handle.
       # This opens us up to letting those object extend themselves naturally.
-      def method_missing(method_sym, *_arguments, &_block)
+      # This will also look for plugins
+      def method_missing(method_sym, *arguments, &_block)
+        # SCM Source
         if AvailableValues.scm.include?(method_sym)
-          # SCM Source
           return env.scm.send(method_sym)
         end
 
+        # Request Source
         if AvailableValues.request_source.include?(method_sym)
-          # Request Source
           return env.request_source.send(method_sym)
         end
 
-        raise "Unknown method '#{method_sym}', please check out the documentation for available variables".red
+        # Plugins
+        class_name = method_sym.to_s.danger_class
+        if Danger::Dangerfile::DSL.const_defined?(class_name)
+          plugin_ref = Danger::Dangerfile::DSL.const_get(class_name)
+          if plugin_ref < Plugin
+            plugin_ref.new(self).run(*arguments)
+          else
+            raise "'#{method_sym}' is not a valid danger plugin".red
+          end
+        else
+          raise "Unknown method '#{method_sym}', please check out the documentation for available plugins".red
+        end
+      end
+
+      private
+
+      def load_default_plugins
+        Dir["./lib/danger/plugins/*.rb"].each do |file|
+          require File.expand_path(file)
+        end
+
+        Dir["./danger_plugins/*.rb"].each do |file|
+          require File.expand_path(file)
+        end
       end
     end
   end